DSS Assertions

Helper functions for DSS type and format checking.

Usage

assert_dss_connected()

assert_dss_file(file)

assert_path_format(path)

assert_parts_format(parts)

assert_read_support(obj)

assert_write_support(obj)

assert_timeseries(obj)

assert_timeseries_type(type)

assert_timeseries_interval(type)

assert_paired(obj)

assert_grid(obj)

assert_attributes(obj, attributes)

Functions

assert_dss_connected(): Assert that DSS dependencies have been loaded.

assert_dss_file(): Assert the supplied object is a DSS file handle.

assert_path_format(): Assert the supplied path is in the correct format.

assert_parts_format(): Assert set of supplied parts is named and complete.

assert_read_support(): Assert the supplied rJava object is a DSS timeseries.

assert_write_support(): Assert the supplied rJava object is a DSS timeseries.

assert_timeseries(): Assert the supplied rJava object is a DSS timeseries.

assert_timeseries_type(): Assert the supplied time series type is recognized.

assert_timeseries_interval(): Assert the supplied time series interval is recognized.

assert_paired(): Assert the supplied rJava object is a DSS paired data container.

assert_grid(): Assert the supplied rJava object is a DSS grid data container.

assert_attributes(): Assert the supplied attributes match the requirements for the supplied rJava object.