Normalize DSS File Path

Normalize a DSS file path. A wrapper for base::normalizePath() with winslash set to "/".

Usage

normalize_path(filename, exists)

Arguments

filename

A DSS file path.

exists

If TRUE, assert that the DSS file exists.

Value

A normalized file path.

Details

This function attempts to strictly resolve file paths by traversing the file path back to the deepest existing directory, in order to resolve potential issues with directory names (e.g., differences in letter case).